requirements.txt
telethon
matplotlib
numpy
python-dotenv
flask